Remaining fix for building in ARM of krita, mostly changes double -> qreal.
svn path=/trunk/koffice/; revision=1151822

+++ krita/plugins/extensions/dockers/colorselectorng/kis_color_selector_ring.cpp
@@ -148,7 +148,7 @@ void KisColorSelectorRing::paintCache()
                         coef+=currentRadius;
                         coef-=innerRadius();
                     }
-                    coef=qBound(0., coef, 1.);
+                    coef=qBound(qreal(0.), coef, qreal(1.));
                     int red=qRed(m_cachedColors.at(angle));
                     int green=qGreen(m_cachedColors.at(angle));
                     int blue=qBlue(m_cachedColors.at(angle));
