+++ krita/plugins/paintops/softbrush/kis_soft_paintop.cpp
@@ -92,8 +92,8 @@ KisSoftPaintOp::KisSoftPaintOp(const KisSoftPaintOpSettings *settings, KisPainte
     m_color = painter->paintColor();
 
     // compute spacing for brush
-    m_xSpacing = qMax(0.5,m_sizeProperties.spacing * m_sizeProperties.diameter * m_sizeProperties.scale);
-    m_ySpacing = qMax(0.5,m_sizeProperties.spacing * m_sizeProperties.diameter * m_sizeProperties.aspect * m_sizeProperties.scale);
+    m_xSpacing = qMax(qreal(0.5),m_sizeProperties.spacing * m_sizeProperties.diameter * m_sizeProperties.scale);
+    m_ySpacing = qMax(qreal(0.5),m_sizeProperties.spacing * m_sizeProperties.diameter * m_sizeProperties.aspect * m_sizeProperties.scale);
     m_spacing = qMax(m_xSpacing, m_ySpacing);
     
 #ifdef BENCHMARK
